Support batched writes (#1049)
## Description

Added a new `BATCH_INSERT_SIZE` parameter that controls the number of
rows to batch while performing an insert.
We were previously limited to `256 Parameters` which was not very
performant for large numbers of writes.

By calling the `interpolateParameters` method, (which is flagged off by
supportManyParameters), we are able to write larger batches of records.
Setting the default value to 1000 rows is meant to be conservative and
should prevent users from hitting the max payload size.

```
Error : com.databricks.internal.sdk.core.error.platform.InvalidParameterValue: The statement field must not exceed a length of 16777216 bytes.
```

In our testing, we discovered that timestamps were not properly
converted by the `interpolateSQL` method, so that logic been
patched/unit tested as well.

## Testing

Additional unit tests that verify the batching of records.
Local behavior verification and - writing to databricks via jdbc.

## Additional Notes to the Reviewer

You currently need the following parameters to enable this behavior:

`supportManyParameters=1` + `EnableBatchedInserts=1` +
`BatchInsertSize=5000`

There is an existing `MAX_BATCH_SIZE` which is used by similar code, but
I think they are separate use cases that warrant different parameters.

Related Issue
https://github.com/databricks/databricks-jdbc/issues/1047

---------

Signed-off-by: sshpuntoff <samshpuntoff@gmail.com>
Co-authored-by: Jose Soto <josecsmorales@gmail.com>